agentic_ai/langgraph_agentic_examples/requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
pandas
